Prettier MD5 checks
Display the progress bar when performing MD5 checks, letting the user
know what's going on after pressing 'U' to upload files. Also, if
verbose, display the current working directory first and then just the
file names, to not clutter the terminal, and inform the user about
skipped files as soon as they're checked against the server.